+++ core/api/typescript/src/lib/enclaves/enclave_context.ts
@@ -556,8 +556,8 @@ export class EnclaveContext {
     }
 
     // Docs available at https://docs.kurtosistech.com/kurtosis-core/lib-documentation
-    public async storeFilesFromUserService(serviceId: ServiceID, absoluteFilepathInUserService: string): Promise<Result<FilesArtifactID, Error>> {
-        const args = newStoreFilesArtifactFromServiceArgs(serviceId, absoluteFilepathInUserService)
+    public async storeFilesFromService(serviceId: ServiceID, absoluteFilepathOnServiceContainer: string): Promise<Result<FilesArtifactID, Error>> {
+        const args = newStoreFilesArtifactFromServiceArgs(serviceId, absoluteFilepathOnServiceContainer)
         const storeFilesArtifactFromServiceResponseResult = await this.backend.storeFilesArtifactFromService(args)
         if (storeFilesArtifactFromServiceResponseResult.isErr()) {
             return err(storeFilesArtifactFromServiceResponseResult.error)